+++ src/components/simplesearchengine/SimpleSearchEngineSecondary.java
@@ -1,0 +1,15 @@
+package components.simplesearchengine;
+
+/**
+ * {@code SimpleSearchEngineKernel<T>} enhanced with secondary methods.
+ *
+ * @param <T>
+ *                type of {@code SimpleSearchEngine} entries
+ *
+ * @author Saurav Poudyel
+ *
+ */
+public abstract class SimpleSearchEngineSecondary<T>
+                implements SimpleSearchEngine<T> {
+
+}